+++ Packs/Code42/Integrations/Code42/CHANGELOG.md
@@ -7,6 +7,11 @@ Added new commands:
         Optionally takes a list of risk tags and only gets employees who have those risk tags.
     - **code42-highriskemployee-add-risk-tags** that takes a username and risk tags and associates the risk tags with the user.
     - **code42-highriskemployee-remove-risk-tags** that takes a username and risk tags and disassociates the risk tags from the user.
+    - **code42-user-deactivate** that deactivates a user in Code42.
+    - **code42-user-reactivate** that reactivates a user in Code42.
+    - **code42-user-block** that blocks a user in Code42.
+    - **code42-user-unblock** that unblocks a user in Code42.
+    - **code42-user-create** that creates a user in Code42.
 Improve error messages for all Commands to include exception detail.
 
 ## [20.3.3] - 2020-03-18